Fix clean shutdown with docker kill
By using "docker kill" to stop the core it will bluntly stop the core. A
graceful shutdown will be be performed because it sends SIGKILL which
cannot be trapped.

The "docker kill --signal SIGINT" command should be used instead, but
this requires that the core is root process in the image. The run.sh
shell script process has to be replaced by the core process, otherwise
the SIGINT will not reach the core.