Add support for SRTv4 clients
Older clients (e.g. Makito encoders) don't support streamid and
couldn't send a stream. This adds support for v4 clients. Such
clients will always be publishing to the resource equal to the
client address, e.g. 192.168.1.42:63793

Clients that want to play this stream are required to send a
streamid with the respective resource name.